+++ doc/whats_new.rst
@@ -64,8 +64,8 @@ Enhancements
      (`#7723 <https://github.com/scikit-learn/scikit-learn/pull/7723>`_)
      by `Mikhail Korobov`_.
 
-   - For sparse matrices, :func:`preprocessing.normalize with ``return_norm=True``
-     will now raise a NotImplementedError with 'l1' or 'l2' norm and with norm 'max'
+   - For sparse matrices, :func:`preprocessing.normalize` with ``return_norm=True``
+     will now raise a ``NotImplementedError`` with 'l1' or 'l2' norm and with norm 'max'
      the norms returned will be the same as for dense matrices (:issue:`7771`).
      By `Ang Lu <https://github.com/luang008>`_.